Da finire ma ho aggiunto cose da fare solo il 16 (add e remove praticamente)

+++ V1/src/ingsoft/commands/running/RemoveCommand.java
@@ -3,11 +3,13 @@ package ingsoft.commands.running;
 import ingsoft.App;
 import ingsoft.ViewSE;
 import ingsoft.commands.AbstractCommand;
+import ingsoft.util.AssertionControl;
 import ingsoft.util.StringUtils;
 
 public class RemoveCommand extends AbstractCommand {
 
     private final App app;
+    private final String NOMECLASSE = this.getClass().getSimpleName();
 
     public RemoveCommand(App app) {
         this.app = app;
@@ -35,9 +37,24 @@ public class RemoveCommand extends AbstractCommand {
         switch (option) {
             case 'c' -> removeConfiguratore(ar);
             case 'f' -> removeFruitore(ar);
-            case 'v' -> {if(app.date.getGiorno() == 16) removeVolontario(ar); else ViewSE.println("Azione possibile solo il 16 del mese!");}
-            case 't' -> {if(app.date.getGiorno() == 16) removeTipoVisita(ar); else ViewSE.println("Azione possibile solo il 16 del mese!");}
-            case 'L' -> {if(app.date.getGiorno() == 16) removeLuogo(ar); else ViewSE.println("Azione possibile solo il 16 del mese!");}
+            case 'v' -> {
+                if (app.date.getGiorno() == 16)
+                    removeVolontario(ar);
+                else
+                    ViewSE.println("Azione possibile solo il 16 del mese!");
+            }
+            case 't' -> {
+                if (app.date.getGiorno() == 16)
+                    removeTipoVisita(ar);
+                else
+                    ViewSE.println("Azione possibile solo il 16 del mese!");
+            }
+            case 'L' -> {
+                if (app.date.getGiorno() == 16)
+                    removeLuogo(ar);
+                else
+                    ViewSE.println("Azione possibile solo il 16 del mese!");
+            }
             default -> ViewSE.println("Opzione non riconosciuta per 'remove'.");
         }
     }
@@ -53,16 +70,34 @@ public class RemoveCommand extends AbstractCommand {
     }
 
     private void removeVolontario(String[] args) {
+        if(!app.alreadyDone16){
+            AssertionControl.logMessage("Non puoi rimuovere un volontario se non è il 16 del mese: " + args[0], 1,
+            NOMECLASSE);
+            return;
+        }
+
         ViewSE.println("Eseguo: Rimuovo volontario");
         app.db.removeVolontario(args[0]);
     }
 
     private void removeTipoVisita(String[] args) {
+        if(!app.alreadyDone16){
+            AssertionControl.logMessage("Non puoi rimuovere un tipo di visita se non è il 16 del mese: " + args[0], 1,
+            NOMECLASSE);
+            return;
+        }
+
         ViewSE.println("Eseguo: Rimuovo tipo visita");
         app.db.removeTipoVisita(args[0]);
     }
 
     private void removeLuogo(String[] args) {
+        if(!app.alreadyDone16){
+            AssertionControl.logMessage("Non puoi rimuovere un luogo se non è il 16 del mese: " + args[0], 1,
+            NOMECLASSE);
+            return;
+        }
+
         ViewSE.println("Eseguo: Rimuovo luogo");
         app.db.removeLuogo(args[0]);
     }
